app_bot/requirements.txt
boto3==1.28.0
coverage==7.2.7
pytest==7.3.1
pytest-mock==3.11.1
pytest-cov
selenium
requests==2.31.0
mock==5.1.0
ffmpeg-python==0.2.0
Flask==2.3.2
Werkzeug==2.2.3
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-api-python-client==2.99.0
msal==1.24.0
